refactor(core): extract activation path reconstruction
