+++ packages/remnic-core/src/graph.ts
@@ -16,6 +16,8 @@ import * as path from "node:path";
 import { readEdgeConfidence } from "./graph-edge-reinforcement.js";
 import { emitGraphEvent } from "./graph-events.js";
 import type { GraphConstructionCapabilitySet } from "./capabilities.js";
+import { reconstructActivationPath, type ActivationPredecessor } from "./graph-path-reconstruction.js";
+export { reconstructActivationPath, type ActivationPredecessor };
 
 export type GraphType = "entity" | "time" | "causal";
 
@@ -401,44 +403,6 @@ interface GraphFileMeta {
   size: number;
   mtimeMs: number;
   ino: number;
-}
-
-export interface ActivationPredecessor {
-  prev: string;
-  edgeConfidence: number;
-  graphType: GraphType;
-}
-
-export function reconstructActivationPath(
-  seed: string,
-  candidate: string,
-  predecessors: Map<string, ActivationPredecessor>,
-  maxSteps: number,
-): ActivationPath | null {
-  const nodeIds = [candidate];
-  const edgeConfidences: number[] = [];
-  const graphTypes: GraphType[] = [];
-  let current = candidate;
-  const stepCap = Number.isFinite(maxSteps)
-    ? Math.max(0, Math.ceil(maxSteps))
-    : predecessors.size + 1;
-  for (let step = 0; step < stepCap && current !== seed; step += 1) {
-    const predecessor = predecessors.get(`${seed}\0${current}`);
-    if (!predecessor) return null;
-    nodeIds.push(predecessor.prev);
-    edgeConfidences.push(predecessor.edgeConfidence);
-    graphTypes.push(predecessor.graphType);
-    current = predecessor.prev;
-  }
-
-  if (current !== seed) return null;
-  nodeIds.reverse();
-  edgeConfidences.reverse();
-  graphTypes.reverse();
-  if (edgeConfidences.length !== nodeIds.length - 1 || graphTypes.length !== nodeIds.length - 1) {
-    return null;
-  }
-  return { nodeIds, edgeConfidences, graphTypes };
 }
 
 export class GraphIndex {
